refactor: reorganize schema types
