+++ crates/tenx-mcp/src/schema/mod.rs
@@ -1,7 +1,28 @@
 pub mod jsonrpc;
-pub mod types;
 pub mod requests;
 
+mod initialization;
+mod capabilities;
+mod implementation;
+mod resources;
+mod prompts;
+mod content;
+mod tools;
+mod sampling;
+mod completions;
+mod roots;
+mod logging;
+
 pub use jsonrpc::*;
-pub use types::*;
 pub use requests::*;
+pub use initialization::*;
+pub use capabilities::*;
+pub use implementation::*;
+pub use resources::*;
+pub use prompts::*;
+pub use content::*;
+pub use tools::*;
+pub use sampling::*;
+pub use completions::*;
+pub use roots::*;
+pub use logging::*;
