+++ crates/tenx-mcp/src/schema/prompts.rs
@@ -1,0 +1,71 @@
+use std::collections::HashMap;
+use super::*;
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct ListPromptsResult {
+    pub prompts: Vec<Prompt>,
+    #[serde(rename = "nextCursor", skip_serializing_if = "Option::is_none")]
+    pub next_cursor: Option<Cursor>,
+}
+
+impl ListPromptsResult {
+    pub fn new() -> Self {
+        Self { prompts: Vec::new(), next_cursor: None }
+    }
+
+    pub fn with_prompt(mut self, prompt: Prompt) -> Self {
+        self.prompts.push(prompt);
+        self
+    }
+
+    pub fn with_prompts(mut self, prompts: impl IntoIterator<Item = Prompt>) -> Self {
+        self.prompts.extend(prompts);
+        self
+    }
+
+    pub fn with_cursor(mut self, cursor: impl Into<Cursor>) -> Self {
+        self.next_cursor = Some(cursor.into());
+        self
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GetPromptResult {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    pub messages: Vec<PromptMessage>,
+    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
+    pub meta: Option<HashMap<String, Value>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Prompt {
+    pub name: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub title: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub arguments: Option<Vec<PromptArgument>>,
+    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
+    pub _meta: Option<HashMap<String, Value>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PromptArgument {
+    pub name: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub title: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub required: Option<bool>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PromptMessage {
+    pub role: Role,
+    pub content: Content,
+}
